Move tests to separate folders, update deep-theory refs to meta-theory
- Rust: extract tests from lib.rs to tests/rml_tests.rs (integration tests)
- JS: rename test/ to tests/ for consistency with Rust structure
- Replace all deep-theory references with meta-theory across docs
- Fix stale test count in README
